requirements.txt
blinker==1.4
bottle==0.12.19
click==8.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pygame==1.9.6
PyJWT==2.3.0
PyMySQL==1.0.2
uWSGI==2.0.20
visitor==0.1.3
Werkzeug==2.0.2
